use CC crate to build dependencies rather than bindgen
