+++ Cargo.toml
@@ -11,7 +11,9 @@ name = "qrllib"
 path = "src/lib.rs"
 
 [dependencies]
+libc = "0.2.0"
 
 [build-dependencies]
+cc = { version = "1.0", features = ["parallel"] }
 cmake = "0.1.46"
 bindgen = "0.59.2"
