+++ src/rustwrapper/dilithium/dilithium_wrapper.c
@@ -1,0 +1,1 @@
+#include "dilithium_wrapper.h"
