+++ src/rustwrapper/dilithium/dilithium_wrapper.h
@@ -1,0 +1,13 @@
+// Distributed under the MIT software license, see the accompanying
+// file LICENSE or http://www.opensource.org/licenses/mit-license.php.
+#ifndef QRLLIB_DILITHIUM_WRAPPER_H
+#define QRLLIB_DILITHIUM_WRAPPER_H
+
+#include <api.h>
+#include <randombytes.h>
+
+const uint32_t WRAPPED_CRYPTO_PUBLICKEYBYTES = CRYPTO_PUBLICKEYBYTES;
+const uint32_t WRAPPED_CRYPTO_SECRETKEYBYTES = CRYPTO_SECRETKEYBYTES;
+const uint32_t WRAPPED_CRYPTO_BYTES = CRYPTO_BYTES;
+
+#endif
