+++ src/rustwrapper/kyber/kyber_wrapper.c
@@ -1,4 +1,3 @@
 // Distributed under the MIT software license, see the accompanying
 // file LICENSE or http://www.opensource.org/licenses/mit-license.php.
-#include <dilithium/ref/api.h>
-#include <dilithium/ref/randombytes.h>
+#include "kyber_wrapper.h"
