+++ src/rustwrapper/kyber/kyber_wrapper.h
@@ -1,0 +1,15 @@
+// Distributed under the MIT software license, see the accompanying
+// file LICENSE or http://www.opensource.org/licenses/mit-license.php.
+#ifndef QRLLIB_KYBER_WRAPPER_H
+#define QRLLIB_KYBER_WRAPPER_H
+
+#define KYBER_K 3
+
+#include <api.h>
+#include <randombytes.h>
+
+const uint32_t WRAPPED_KYBER_PUBLICKEYBYTES = KYBER_PUBLICKEYBYTES;
+const uint32_t WRAPPED_KYBER_SECRETKEYBYTES = KYBER_SECRETKEYBYTES;
+const uint32_t WRAPPED_KYBER_CIPHERTEXTBYTES = KYBER_CIPHERTEXTBYTES;
+
+#endif
