+++ src/rustwrapper/kyber/mod.rs
@@ -1,0 +1,1 @@
+pub mod kyber;
