--- src/rustwrapper/kyber_wrapper.hxx
@@ -1,5 +1,0 @@
-// Distributed under the MIT software license, see the accompanying
-// file LICENSE or http://www.opensource.org/licenses/mit-license.php.
-#define KYBER_K 3
-#include <kyber/ref/api.h>
-#include <kyber/ref/randombytes.h>
