+++ tests/rust/dilithium/dilithium_test.rs
@@ -1,6 +1,6 @@
 // Distributed under the MIT software license, see the accompanying
 // file LICENSE or http://www.opensource.org/licenses/mit-license.php.
-use qrllib::rustwrapper::dilithium::Dilithium;
+use qrllib::rustwrapper::dilithium::dilithium::Dilithium;
 
 #[test]
 fn sign_keypair() {
