+++ tests/rust/dilithium/mod.rs
@@ -1,0 +1,2 @@
+mod dilithium_ref_test;
+mod dilithium_test;
