+++ tests/rust/kyber/mod.rs
@@ -1,0 +1,2 @@
+mod kyber_ref_test;
+mod kyber_test;
