+++ tests/rust/mod.rs
@@ -1,6 +1,4 @@
 // Distributed under the MIT software license, see the accompanying
 // file LICENSE or http://www.opensource.org/licenses/mit-license.php.
-mod dilithium_ref_test;
-mod dilithium_test;
-mod kyber_ref_test;
-mod kyber_test;
+mod dilithium;
+mod kyber;
